+++ src/extension.ts
@@ -8,6 +8,7 @@ import { ViewType, ContextName } from './constants';
 import { registerTelemetryReporter, sendEvent } from './util/telemetry';
 import { registerScanReferencesCommand } from './util/dependency';
 import { locale } from './util/i18n';
+import { registerHoiFs } from './util/hoifs';
 
 export function activate(context: vscode.ExtensionContext) {
     // Must register this first because other component may use it.
@@ -20,6 +21,7 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(registerModFile());
     context.subscriptions.push(worldMap.register());
     context.subscriptions.push(registerScanReferencesCommand());
+    context.subscriptions.push(registerHoiFs());
 
     // Use proposed vscode API
     context.subscriptions.push(vscode.window.registerCustomEditorProvider(ViewType.DDS, new DDSViewProvider()));